# docs/openapi.yaml
openapi: 3.1.1
info:
  title: AgentFeedback API
  version: '1.0'
  license:
    name: MIT
    url: https://opensource.org/license/mit
  description: >-
    HTTP API of AgentFeedback, a service that stores write-once feedback from AI
    coding agents. The contract is this document together with the JSON Schemas
    under `schemas/` and the conformance kit under `conformance/`. This is the API
    of the next major release. Schemas use the custom keyword `x-max-bytes` for
    UTF-8 byte limits, since JSON Schema `maxLength` counts code points. The
    server serves this document at `/api/v1/openapi.json` in bundled form, with
    the referenced schema files inlined.
servers:
  - url: https://api.agentfeedback.io
    description: Hosted
  - url: http://127.0.0.1:8090
    description: Local
security:
  - bearer: []
  - apiKey: []
...
paths:
...
  /api/v1/schemas:
    get:
      tags: [schemas]
      operationId: listSchemas
      summary: List known schemas
      description: Kinds and versions the server knows, including `envelope`. Unauthenticated; carries no data.
      security: []
      responses:
        '200':
          description: Known schemas.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SchemasList'}
              example:
                schemas:
                  - {kind: envelope, versions: [1]}
                  - {kind: friction, versions: [1]}
                  - {kind: review, versions: [1]}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '500': {$ref: '#/components/responses/InternalError'}
...
  /api/v1/openapi.json:
    get:
      tags: [discovery]
      operationId: getOpenApi
      summary: This OpenAPI document
      description: The OpenAPI document as JSON. Unauthenticated; carries no data.
      security: []
      responses:
        '200':
          description: The OpenAPI document.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
          content:
            application/json:
              schema: {type: object}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '500': {$ref: '#/components/responses/InternalError'}
  /mcp:
    post:
      tags: [mcp]
      operationId: postMcp
      summary: Remote MCP
      description: >-
        Remote MCP server over the MCP Streamable HTTP transport. Authenticated
        like `/api/v1/*`.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/JsonRpc'}
      responses:
        '200':
          $ref: '#/components/responses/JsonRpc'
        '202':
          description: 'Accepted: the body was a JSON-RPC notification or response; no body is returned.'
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
            Mcp-Session-Id: {$ref: '#/components/headers/McpSessionId'}
        '400': {$ref: '#/components/responses/McpBadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/McpNotFound'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/McpTooLarge'}
        '415': {$ref: '#/components/responses/McpUnsupportedMediaType'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
  /skill:
    get:
      tags: [discovery]
      operationId: getSkill
      summary: Submission guidance
      description: >-
        The rendered submission guidance with this server's base URL
        substituted. Unauthenticated; carries no data.
      security: []
      parameters:
        - name: format
          in: query
          description: >-
            `skill-md` (default, runs the binary), `agents-md` or `prompt`
            (HTTP instructions inline).
          schema: {type: string, enum: [skill-md, agents-md, prompt], default: skill-md}
      responses:
        '200':
          description: Guidance as Markdown.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
          content:
            text/markdown:
              schema: {type: string}
        '400': {$ref: '#/components/responses/BadRequest'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '500': {$ref: '#/components/responses/InternalError'}
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <key>`; scheme name case-insensitive. Either header authorises; neither shadows the other.'
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: '`X-Api-Key: <key>`. Either header authorises; neither shadows the other.'
  headers:
    XRequestId:
      description: Request identifier; present on every response.
      schema: {type: string}
    CacheControl:
      description: '`no-store` on every response of an authenticated route, success or error.'
      schema: {type: string, const: no-store}
    WWWAuthenticate:
      description: '`Bearer` on 401.'
      schema: {type: string}
...
    Allow:
      description: Methods the route accepts.
      schema: {type: string}
    McpSessionId:
      description: Session identifier assigned at initialization; the client sends it back on later requests.
      required: false
      schema: {type: string}
    RetryAfter:
      description: Seconds to wait before retrying, when known.
      schema: {type: integer, minimum: 0}
...
  responses:
...
    JsonRpc:
      description: JSON-RPC 2.0 response (MCP Streamable HTTP transport).
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
        Mcp-Session-Id: {$ref: '#/components/headers/McpSessionId'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/JsonRpc'}
        text/event-stream:
          schema:
            type: string
            description: >-
              The server may answer a request with a server-sent event stream
              carrying JSON-RPC messages when the client's `Accept` header
              allows it.
    McpBadRequest:
      description: >-
        On protocol 2026-07-28 and later, a JSON-RPC error response when the
        MCP transport requires one (an unsupported protocol version in
        `_meta`, invalid params, an unknown tool). Otherwise the Error shape:
        `bad_request` (a method the server does not handle on an earlier
        protocol version, a protocol version header the server does not
        support, an `Accept` header without both `application/json` and
        `text/event-stream`, a body that is not a JSON-RPC message) or
        `validation_error` (a `{project}` that is empty once normalised).
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/Error'
              - $ref: '#/components/schemas/JsonRpcError'
    McpNotFound:
      description: >-
        JSON-RPC error `-32601` for a method the server does not handle, on
        protocol 2026-07-28 and later. Earlier protocol versions get 400
        `bad_request` in the Error shape instead.
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/JsonRpcError'}
    McpTooLarge:
      description: >-
        `request_too_large`: the request body is over 10551296 bytes, the
        create body limit (10485760) plus 65536 bytes for the JSON-RPC
        message. Tool arguments over the create or mark body limit get a tool
        result with `isError` and the same error body instead.
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    McpUnsupportedMediaType:
      description: '`bad_request`: the request Content-Type is not `application/json`.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    BadRequest:
      description: >-
        `bad_request` (body not JSON, or not an object) or `validation_error`
        (reads, processing, import, path parameters).
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example:
            error: validation_error
            message: limit must be between 1 and 500
            request_id: 3f9c2a7e5b1d4c08
            details:
              - code: out_of_range
                pointer: '?limit'
                message: limit must be between 1 and 500
    Unauthorized:
      description: '`unauthorized`.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
        WWW-Authenticate: {$ref: '#/components/headers/WWWAuthenticate'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
    MethodNotAllowed:
      description: '`method_not_allowed`.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
        Allow: {$ref: '#/components/headers/Allow'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
    RateLimited:
      description: '`rate_limited` (hosted service only).'
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    InternalError:
      description: '`internal_error`.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Unavailable:
      description: '`unavailable`; `Retry-After` when known.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
...
    Warning:
      type: object
      additionalProperties: false
      required: [code, pointer, message]
      properties:
        code: {type: string}
        pointer: {type: string, description: 'RFC 6901 pointer into the stored record; query parameters use `?name`.'}
        message: {type: string}
        existing_id: {type: integer, minimum: 1, description: Only on the 409 `key_reused` detail.}
    Error:
      type: object
      required: [error, message, request_id]
      properties:
        error:
          type: string
          enum: [bad_request, validation_error, unauthorized, not_found, method_not_allowed, replay_mismatch, request_too_large, rate_limited, internal_error, unavailable]
        message: {type: string, description: 'Names the field or parameter, the value received when short, and the accepted range or shape.'}
        request_id: {type: string}
        details:
          type: array
          items: {$ref: '#/components/schemas/Warning'}
...
    KindVersions:
      type: object
      required: [kind, versions]
      properties:
        kind: {type: string}
        versions:
          type: array
          items: {type: integer, minimum: 1}
    SchemasList:
      type: object
      required: [schemas]
      properties:
        schemas:
          type: array
          items: {$ref: '#/components/schemas/KindVersions'}
...
    JsonRpc:
      type: object
      description: A JSON-RPC 2.0 message as defined by the MCP Streamable HTTP transport.
    JsonRpcError:
      type: object
      description: A JSON-RPC 2.0 error response.
      required: [jsonrpc, error]
      properties:
        jsonrpc: {type: string, const: '2.0'}
        id:
          oneOf:
            - {type: string}
            - {type: integer}
            - {type: 'null'}
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: integer}
            message: {type: string}
            data: {}